Config load refactored
BuildingModifier
