--- Source/BuildingModifierMod/IOHelper.cs
@@ -1,15 +1,0 @@
-﻿namespace BuildingModifierMod
-{
-    using System.IO;
-
-    public static class IOHelper
-    {
-        public static void EnsureDirectoryExists(string path)
-        {
-            if (!Directory.Exists(path))
-            {
-                Directory.CreateDirectory(path);
-            }
-        }
-    }
-}
