+++ Source/BuildingModifierMod/JsonFileManager.cs.bak
